sub_realtime_translator/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
cffi==1.17.1
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
numpy==2.3.1
propcache==0.3.2
pycparser==2.22
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dotenv==1.1.1
shiboken6==6.9.1
sounddevice==0.5.2
websockets==15.0.1
yarl==1.20.1
